Refactor into submodule
