+++ crates/esthri/src/ops/mod.rs
@@ -13,6 +13,5 @@
 pub(super) mod copy;
 pub(super) mod delete;
 pub(super) mod download;
-pub(super) mod presign;
 pub(super) mod sync;
 pub(super) mod upload;
